Improved rotten tomatoes search, organized files, and hooked up amazon/netflix
