+++ extract/tv_by_the_numbers.py
@@ -1,3 +1,6 @@
+import sys
+sys.path.append('..')
+
 import requests
 from bs4 import BeautifulSoup
 import pandas as pd
